requirements.txt
asgiref==3.10.0
Django==5.2.7
django-appconf==1.1.0
django-ckeditor==6.7.3
django-imagekit==6.0.0
django-js-asset==3.1.2
pilkit==3.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
whitenoise==6.11.0
